Reswizzle and restructure components and utils
