+++ src/theme/DocSidebar/Desktop/Content/index.tsx
@@ -1,17 +1,16 @@
 /**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * SWIZZLED VERSION: 2.0.0-rc.1
+ * REASONS:
+ *  - Add version picker.
  */
 
 import React, { useState } from 'react';
 import clsx from 'clsx';
+import { ThemeClassNames } from '@docusaurus/theme-common';
 import {
-  ThemeClassNames,
   useAnnouncementBar,
   useScrollPosition,
-} from '@docusaurus/theme-common';
+} from '@docusaurus/theme-common/internal';
 import DocSidebarItems from '@theme/DocSidebarItems';
 import type { Props } from '@theme/DocSidebar/Desktop/Content';
 import VersionPicker from '@site/src/components/VersionPicker';
