+++ src/theme/Footer/index.tsx
@@ -1,36 +1,22 @@
-import React from 'react';
-
-import { useThemeConfig } from '@docusaurus/theme-common';
-import FooterLinks from '@theme/Footer/Links';
-import FooterLogo from '@theme/Footer/Logo';
-import FooterCopyright from '@theme/Footer/Copyright';
-import FooterLayout from '@theme/Footer/Layout';
-import Social from '@site/src/components/Social';
-
 /**
- * SWIZZLED VERSION: 2.0.0-beta.20
+ * SWIZZLED VERSION: 2.0.0-rc.1
  * REASONS:
- * - The socials bar needed full width, so needed to be placed outside the default container.
+ *  - The socials bar needed full width, so needed to be placed outside the default container.
  */
 
-function Footer(): JSX.Element | null {
-  const { footer } = useThemeConfig();
-  if (!footer) {
-    return null;
-  }
-  const { copyright, links, logo, style } = footer;
+import React from 'react';
+import Footer from '@theme-original/Footer';
+import type FooterType from '@theme/Footer';
+import type { WrapperProps } from '@docusaurus/types';
+import Social from '@site/src/components/Social';
 
+type Props = WrapperProps<typeof FooterType>;
+
+export default function FooterWrapper(props: Props): JSX.Element {
   return (
     <>
-      <FooterLayout
-        style={style}
-        links={links && links.length > 0 && <FooterLinks links={links} />}
-        logo={logo && <FooterLogo logo={logo} />}
-        copyright={copyright && <FooterCopyright copyright={copyright} />}
-      />
+      <Footer {...props} />
       <Social />
     </>
   );
 }
-
-export default React.memo(Footer);
